+++ src/Rib/Extra/CSS.hs
@@ -5,6 +5,8 @@
 module Rib.Extra.CSS where
 
 import Clay
+import qualified Data.Text as T
+import Lucid
 import Relude
 
 -- | Stock CSS for the <kbd> element
@@ -25,3 +27,14 @@ mozillaKbdStyle = do
   fontWeight $ weight 700
   lineHeight $ px 1
   whiteSpace nowrap
+
+-- | Include the specified Google Fonts
+googleFonts :: [Text] -> Html ()
+googleFonts fs =
+  let fsEncoded = T.intercalate "|" $ T.replace " " "+" <$> fs
+      fsUrl = "https://fonts.googleapis.com/css?family=" <> fsEncoded <> "&display=swap"
+   in stylesheet fsUrl
+
+-- | Include the specified stylesheet URL
+stylesheet :: Text -> Html ()
+stylesheet x = link_ [rel_ "stylesheet", href_ x]