+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - pandoc-types: 1.20
 - New features:
   - API exposes the CLI parser (`optparse-applicative`) for user-level composition
+  - Add `Rib.Parser.Pandoc.getToC` returning rendered Table of contents for a Pandoc document
+  - `Rib.Extra.CSS`: add `googleFonts` and `stylesheet`
 
 ## 0.7.0.0